refactor: split armor location validation renderers
